+++ main-presentation/src/main/java/com/teamhy2/feature/home/HomeViewModel.kt
@@ -3,7 +3,7 @@ package com.teamhy2.feature.home
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.teamhy2.feature.home.model.HomeUiState
-import com.teamhy2.hongikyeolgong2.timer.prsentation.model.TimerUiModel
+import com.teamhy2.hongikyeolgong2.timer.presentation.model.TimerUiModel
 import com.teamhy2.main.domain.model.WeeklyStudyDay
 import com.teamhy2.main.domain.model.WiseSaying
 import com.teamhy2.main.domain.repository.StudyDayRepository
